Make shaptools available for venv-salt-minion (bsc#1212695)
This commit makes use of RPM Supplements attribute to trigger the
installation of the "python3-shaptools-venv-salt-minion" package when
"venv-salt-minion" and "python3-shaptools" packages are going to be
both installed on the system.

This new "python3-shaptools-venv-salt-minion" just takes care of
creating the necessary symlink to make "shaptools" available inside the
Salt Bundle.